pool: use `Executor` trait instead of concrete type
Alternative to https://github.com/hyperium/hyper-util/pull/24.

Currently, `Pool` cannot be used at all. It requires a concrete `Exec`,
and that is private. The #24 PR simply made `Exec` public. However,
there is already other things in this crate that require an `Exec`
(`legacy::Client`).

This adapts `Pool` to follow the same pattern as used for `Client`,
taking in a trait.

This change is NOT a breaking change, as `Pool` cannot be used at all
currently, so no one could be broken.