refactor: run dns resolution in the same tracing-span as the caller (#134)
This makes it possible to trace the entire request, including DNS
resolution.

The use case for this is to be able to suppress specific requests from
being traced in a situation where the request is used to transmit logs
to a remote server. This would result in an infinite loop of logs being
sent to the remote server. https://github.com/tokio-rs/tokio/issues/6659
has more details.