client: support custom pool keys
This PR introduces the ability to use a custom pool key with the legacy
Client.

https://github.com/hyperium/hyper-util/pull/16 added generic support to
`Pool` itself long ago, but it was never fully finished to allow a
custom key in `Client`.

To support backwards compatibility, a default PoolKey is assigned to the
generic parameter which seems reasonably backwards compatible(?)

When providing a custom pool key, the user is required to also pass a
constructor that generates the pool key from the `http::request::Parts`.
I had also considered making a user pass in the PoolKey as part of the
`request()` call, but I think this is a worse experience generally and
harder to be backwards compatible. If a user did want to do a
per-request key they can always set an extension in `Parts` so this
approach is equally flexible.