feat(pool): add a Cache pooling service (#227)
This creates a new `Cache` type that acts as a pool over an inner make-service. Asking the Cache to produce a service will either return a cached one, or create a new one with the inner maker. The returned service is `Cached<S>`, which knows to return the service to the cache upon Drop.

Closes https://github.com/hyperium/hyper/issues/3954