feat(pool): add a Negotiate pooling service (#228)
Creates a `Negotiate` type that uses an inner make-service, inspects the result, and then either returns an upgraded service, or fallback, depending on the negotiated. This is somewhat like a `Router`, but instead of routing on the request, it's routing on the response. It's an advanced form of an `Either`. The original use case is that of a connector that can sends ALPN to a destination, and can either comeback with HTTP/2 negotiated, or fallback to HTTP/1.

Closes https://github.com/hyperium/hyper/issues/3955