feat(pool): add a Map pool service type (#247)
The `map` pool isn't a typical Service like the other pools, but rather is a stand-alone type that maps requests to keys and connectors. As a kind of router, and it cannot determine which inner service to check for backpressure until the request is made. The map implementation allows customization of extracting a key, and how to construct a connector for that key.

Closes https://github.com/hyperium/hyper/issues/3979